requirements.txt
streamlit==1.31.0
google-genai
google-generativeai>=0.3.0
pymongo==4.6.1
mysql-connector-python>=8.0.33
python-dotenv==1.0.1
pandas==2.2.0
Pillow==10.0.0
folium==0.20.0
streamlit-folium==0.26.2
requests>=2.31.0
exifread>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
clip-by-openai
ftfy
regex

# Satellite System - Image Processing & Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0

# Satellite System - API & Web Services
rasterio>=1.3.0
shapely>=2.0.0
geopy>=2.3.0

# Satellite System - Scheduling & Automation
schedule>=1.2.0
APScheduler>=3.10.0

# Next.js Integration Backend API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
